Move monitoring to own package
